Re-use analysis context between executions of AnalysisImpl.
This allows us to take advantage of the analysis cache so that we
don't do redundant analysis when multiple input files import (or
export) common libraries.

This saves a tremendous amount of time when analyzing complex
projects.  For example, we can now analyze all of "pub" in about 1.5
minutes instead of 20 minutes.