Fix .packages file discovery on Windows.
The _discoverPackagespec() method was trying to convert its "root"
argument to a directory URI (one ending in a "/") by calling "new
Uri.directory(root.path)".  This was causing confusion on Windows,
since root.path is in posix-style but the Uri.directory constructor
expects a native OS path.

Fortunately, the caller always passes a directory URI anyhow, so no
conversion is necessary.  We can just use the URI as is.